mosquittoAdapter/requirements.txt
paho-mqtt==1.6.1
fastapi==0.95.2
uvicorn==0.22.0
